KAFKA-4180 : Authentication with multiple actives Kafka producers/consumers
Changed caching in LoginManager to allow one LoginManager per client
JAAS configuration.
Introduced abstract class SaslEndToEndAuthorizationTest.
Added test to End2EndAuthorization for SASL Plain and Gssapi with two
consumers with different credentials.
